Move success states into ButtonWithResponse
